feat: init block ui code structrue
