+++ packages/umi-plugin-ui/src/index.js
@@ -1,6 +1,6 @@
 export default api => {
   require('./plugins/dashboard/index').default(api);
-  // require('./plugins/blocks/index').default(api);
   require('./plugins/configuration/index').default(api);
   require('./plugins/tasks/index').default(api);
+  require('./plugins/blocks/index').default(api);
 };
