+++ packages/umi-plugin-ui/src/plugins/blocks/package.json
@@ -1,6 +1,7 @@
 {
   "peerDependencies": {
     "react": "^16.8.6",
-    "react-dom": "^16.8.6"
+    "react-dom": "^16.8.6",
+    "antd": "^4.0.0-alpha.0"
   }
 }
