+++ packages/umi-plugin-ui/src/plugins/blocks/typings.d.ts
@@ -1,0 +1,2 @@
+declare module '*.less';
+declare module '*.css';
